phish-sim/ml/requirements.txt
# SPDX-License-Identifier: MIT
# Copyright (c) 2024 Phish-Sim Project

# Core ML libraries
torch==2.1.0
transformers==4.35.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# NLP and text processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
tokenizers==0.15.0

# Model serving and API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
joblib==1.3.2

# Data processing
datasets==2.14.6
huggingface-hub==0.19.4

# Evaluation and metrics
evaluate==0.4.1
rouge-score==0.1.2
seqeval==1.2.2

# Visualization and monitoring
matplotlib==3.8.2
seaborn==0.13.0
tensorboard==2.15.1
wandb==0.16.0

# Utilities
tqdm==4.66.1
click==8.1.7
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0